pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "danotes"
version = "0.1.0"
description = "CLI for managing .dan format notes"
readme = "README.md"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.8"
dependencies = [
    "pyfiglet>=1.0.3", 
    "pyyaml>=6.0"
]

[project.scripts]
danotes = "danotes.cli:main"
